+++ components/confluence-sync/src/confluence/errors/AttachmentsNotFoundError.ts
@@ -1,10 +1,13 @@
 // SPDX-FileCopyrightText: 2024 Telefónica Innovación Digital
 // SPDX-License-Identifier: Apache-2.0
 
-export class AttachmentsNotFoundError extends Error {
+import { CustomError } from "./CustomError";
+import { getCauseMessage } from "./ErrorHelpers";
+
+export class AttachmentsNotFoundError extends CustomError {
   constructor(id: string, options?: ErrorOptions) {
     super(
-      `Error getting attachments of page with id ${id}: ${options?.cause}`,
+      `Error getting attachments of page with id ${id}: ${getCauseMessage(options?.cause)}`,
       options,
     );
   }