+++ components/confluence-sync/src/confluence/errors/axios/BadRequestError.ts
@@ -1,9 +1,9 @@
 // SPDX-FileCopyrightText: 2024 Telefónica Innovación Digital
 // SPDX-License-Identifier: Apache-2.0
-
+import { CustomError } from "../CustomError";
 import type { AxiosError } from "axios";
 
-export class BadRequestError extends Error {
+export class BadRequestError extends CustomError {
   constructor(error: AxiosError) {
     super(
       `Bad Request