+++ components/confluence-sync/src/confluence/errors/axios/UnexpectedError.ts
@@ -1,7 +1,9 @@
 // SPDX-FileCopyrightText: 2024 Telefónica Innovación Digital
 // SPDX-License-Identifier: Apache-2.0
 
-export class UnexpectedError extends Error {
+import { CustomError } from "../CustomError";
+
+export class UnexpectedError extends CustomError {
   constructor(error: unknown) {
     super(`Unexpected Error: ${error}`);
   }