+++ components/markdown-confluence-sync/jest.unit.config.cjs
@@ -20,10 +20,10 @@ module.exports = {
   // An object that configures minimum threshold enforcement for coverage results
   coverageThreshold: {
     global: {
-      branches: 99,
-      functions: 99,
-      lines: 99,
-      statements: 99,
+      branches: 98,
+      functions: 98,
+      lines: 98,
+      statements: 98,
     },
   },